fix non debug output also if config debug setting is disabled
