+++ tests/exit1.sh
@@ -1,0 +1,3 @@
+#! /bin/bash
+echo "Error message here for exit code = 1"
+exit 1
